requirements.txt
numpy==1.26.4
pandas==2.2.3
scipy==1.15.1
matplotlib==3.10.0
seaborn
scikit-learn
scikit-image
opencv-python==4.11.0.86
pillow
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
transformers
huggingface_hub
monai==1.4.0
albumentations==2.0.5
tqdm
einops==0.8.1
timm==1.0.15